django_project/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0
Pillow==10.0.0
drf-spectacular==0.26.5
django-storages==1.14.2
boto3==1.28.85
azure-storage-blob==12.18.3
